+++ src/Media/Audio/AudioPlayer.h
@@ -9,7 +9,7 @@
 #include "Engine/Spells/SpellEnums.h"
 #include "Engine/Objects/ActorEnums.h"
 
-#include "Media/Media.h"
+#include "Media/AudioTrack.h"
 
 #include "Utility/String.h"
 #include "Utility/Memory/Blob.h"
