+++ src/Media/Audio/SoundInfo.h
@@ -3,9 +3,9 @@
 #include <string>
 #include <memory>
 
-#include "Utility/Memory/Blob.h"
+#include "Media/AudioDataSource.h"
 
-#include "Media/Media.h"
+#include "Utility/Memory/Blob.h"
 
 #include "SoundEnums.h"
