+++ src/Media/MediaPlayer.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <memory>
 
-#include "Media/Media.h"
+#include "Media/Movie.h"
 
 // MOVIE_3DOLogo  "3dologo"
 // MOVIE_NWCLogo  "new world logo"
